requirements.txt
langchain-google-genai>=0.0.5
langchain-core>=0.1.15
streamlit>=1.31.0
python-dotenv>=1.0.0 